code: fixed compatibility issues with trigo functions

+++ src/transformations/rotation.h
@@ -4,6 +4,8 @@
 #include "transformation.h"
 #include "utils/deg2rad.h"
 
+#include <cmath>
+
 /**
  * @class RotationX
  * @brief Represents a rotation transformation around the X-axis.
@@ -12,8 +14,8 @@ class RotationX : public Transformation {
 public:
     RotationX(float angleInDegrees) {
         float angInRad = DEG2RAD * angleInDegrees;
-        float cos = std::cosf(angInRad);
-        float sin = std::sinf(angInRad);
+        float cos = std::cos(angInRad);
+        float sin = std::sin(angInRad);
 
         matrix_ << 
             1, 0, 0, 0,
@@ -31,8 +33,8 @@ class RotationY : public Transformation {
 public:
     RotationY(float angleInDegrees) {
         float angInRad = DEG2RAD * angleInDegrees;
-        float cos = std::cosf(angInRad);
-        float sin = std::sinf(angInRad);
+        float cos = std::cos(angInRad);
+        float sin = std::sin(angInRad);
 
         matrix_ << 
             cos,  0, sin, 0,
@@ -50,8 +52,8 @@ class RotationZ : public Transformation {
 public: 
     RotationZ(float angleInDegrees) {
         float angInRad = DEG2RAD * angleInDegrees;
-        float cos = std::cosf(angInRad);
-        float sin = std::sinf(angInRad);
+        float cos = std::cos(angInRad);
+        float sin = std::sin(angInRad);
 
         matrix_ << 
             cos, -sin, 0, 0,
